2019-11-06
예은,,,,
다시,,,, 날다,,,,,,

+++ src/member/controller/AdMemberDeleteServlet.java
@@ -39,22 +39,23 @@ public class AdMemberDeleteServlet extends HttpServlet {
 				int result1 = 0, result2 = 0, result3 = 0, result4 = 0;
 				String checkBox = request.getParameter("lists");
 				String userid = request.getParameter("userid");
+				String anwser = request.getParameter("anwser");
 				
-				if(checkBox != null) {
+				if(checkBox != null && anwser != null) {
 					// 체크박스 삭제
 					String check[] = checkBox.split(",");
 					
 					for(int i = 0; i < check.length; i++) {
 						if(!check[i].equals("admin00")) {
 						result1 = mService.deleteMember(check[i]);
-						result2 = qService.insertMember(check[i]);
+						result2 = qService.insertMember(check[i], anwser);
 						}
 					}
-				} else if(userid != null) {
+				} else if(userid != null && anwser != null) {
 				// 회원 상세보기에서 삭제
 					if(!userid.equals("admin00")) {
 					result3 = mService.deleteMember(userid);
-					result4 = qService.insertMember(userid);
+					result4 = qService.insertMember(userid, anwser);
 					}
 				}
